The workhorses of this operation arrive.
bashstick.js:
* \slaps roof of script\ This bad boy can fit so much calculation in it.
* Well, only one technically. It checks for the height of h2 (content +
padding), then applies it to the "top" style of all h3s. Neato!
**  It only checks the first in the h2 array which sacrifices pure
compatibility for performance, but let's be real, they should all be
uniform size anyway.
* Of course, first it has to get the go-ahead from...

chklib.js:
* Awww yiss, WB Docs have got their first library! Gettin' all modular
up in here.
* This one is making sure stuff works properly, then shares with other
things what it finds. It's a helpful little guy!
* Right now it's only owning two skills: debugging and telling
bashstick.js over there that it's OK to do the fancy stuff.

So here's how things work around here:
1 > chklib makes sure the user can use the 'sticky' CSS rule. If yes,
bashstick pulls out the ruler, all good! If not, it tells bashstick to
shush.
2 > If for some reason the user has scripts disabled, the CSS has
'@supports' to make a similar check and uses magic numbers to get things
close enough.
STOP > If the browser doesn't support 'sticky' then the page defaults to
the original implementation. Both the CSS and the script will come to
the same conclusion.